Reorganize all modifiers into modules
Move all the modifiers under a top-level `modifier` module, and further
group them based on the phase they modify under three sub-modules
`init`, `update`, and `render`.

Add a `prelude` module exporting most public API types.

Clean-up the `ForceFieldModifier` with various renames for clarity and
consistency, and added documentation.
