+++ examples/billboard.rs
@@ -5,9 +5,9 @@ use bevy::{
     prelude::*,
     render::{camera::Projection, render_resource::WgpuFeatures, settings::WgpuSettings},
 };
-
-use bevy_hanabi::*;
 use bevy_inspector_egui::WorldInspectorPlugin;
+
+use bevy_hanabi::prelude::*;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut options = WgpuSettings::default();
