+++ src/modifier/render.rs
@@ -1,0 +1,62 @@
+//! Modifiers to influence the rendering of each particle.
+
+use bevy::prelude::*;
+
+use crate::{Gradient, RenderLayout};
+
+/// Trait to customize the rendering of alive particles each frame.
+pub trait RenderModifier {
+    /// Apply the modifier to the render layout of the effect instance.
+    fn apply(&self, render_layout: &mut RenderLayout);
+}
+
+/// A modifier modulating each particle's color by sampling a texture.
+#[derive(Default, Clone)]
+pub struct ParticleTextureModifier {
+    /// The texture image to modulate the particle color with.
+    pub texture: Handle<Image>,
+}
+
+impl RenderModifier for ParticleTextureModifier {
+    fn apply(&self, render_layout: &mut RenderLayout) {
+        render_layout.particle_texture = Some(self.texture.clone());
+    }
+}
+
+/// A modifier modulating each particle's color over its lifetime with a
+/// gradient curve.
+#[derive(Default, Clone)]
+pub struct ColorOverLifetimeModifier {
+    /// The color gradient defining the particle color based on its lifetime.
+    pub gradient: Gradient<Vec4>,
+}
+
+impl RenderModifier for ColorOverLifetimeModifier {
+    fn apply(&self, render_layout: &mut RenderLayout) {
+        render_layout.lifetime_color_gradient = Some(self.gradient.clone());
+    }
+}
+
+/// A modifier modulating each particle's size over its lifetime with a gradient
+/// curve.
+#[derive(Default, Clone)]
+pub struct SizeOverLifetimeModifier {
+    /// The size gradient defining the particle size based on its lifetime.
+    pub gradient: Gradient<Vec2>,
+}
+
+impl RenderModifier for SizeOverLifetimeModifier {
+    fn apply(&self, render_layout: &mut RenderLayout) {
+        render_layout.size_color_gradient = Some(self.gradient.clone());
+    }
+}
+
+/// Reorients the vertices to always face the camera when rendering.
+#[derive(Default, Clone, Copy)]
+pub struct BillboardModifier;
+
+impl RenderModifier for BillboardModifier {
+    fn apply(&self, render_layout: &mut RenderLayout) {
+        render_layout.billboard = true;
+    }
+}
