+++ src/modifier/update.rs
@@ -1,0 +1,145 @@
+//! Modifiers to influence the update behavior of particle each frame.
+//!
+//! The update modifiers control how particles are updated each frame by the
+//! update compute shader.
+
+use bevy::prelude::*;
+
+use crate::UpdateLayout;
+
+/// Trait to customize the updating of existing particles each frame.
+pub trait UpdateModifier {
+    /// Apply the modifier to the update layout of the effect instance.
+    fn apply(&self, update_layout: &mut UpdateLayout);
+}
+
+/// A modifier to apply a constant acceleration to all particles each frame.
+/// Used to simulate gravity.
+#[derive(Default, Clone, Copy)]
+pub struct AccelModifier {
+    /// The constant acceleration to apply to all particles in the effect each
+    /// frame.
+    pub accel: Vec3,
+}
+
+impl UpdateModifier for AccelModifier {
+    fn apply(&self, layout: &mut UpdateLayout) {
+        layout.accel = self.accel;
+    }
+}
+
+/// A single attraction or repulsion source of a [`ForceFieldModifier`].
+///
+/// The source applies a radial force field to all particles around its
+/// position, with a decreasing intensity the further away from the source the
+/// particle is. This force is added to the one(s) of all the other active
+/// sources of a [`ForceFieldModifier`].
+#[derive(Debug, Clone, Copy)]
+pub struct ForceFieldSource {
+    /// Position of the source.
+    pub position: Vec3,
+    /// Maximum radius of the sphere of influence, outside of which
+    /// the contribution of this source to the force field is null.
+    pub max_radius: f32,
+    /// Minimum radius of the sphere of influence, inside of which
+    /// the contribution of this source to the force field is null, avoiding the
+    /// singularity at the source position.
+    pub min_radius: f32,
+    /// The intensity of the force of the source is proportional to its mass.
+    /// Note that the update shader will ignore all subsequent force field
+    /// sources after it encountered a source with a mass of zero. To change
+    /// the force from an attracting one to a repulsive one, simply
+    /// set the mass to a negative value.
+    pub mass: f32,
+    /// The source force is proportional to `1 / distance^force_exponent`.
+    pub force_exponent: f32,
+    /// If `true`, the particles which attempt to come closer than
+    /// [`min_radius`] from the source position will conform to a sphere of
+    /// radius [`min_radius`] around the source position, appearing like a
+    /// recharging effect.
+    ///
+    /// [`min_radius`]: ForceFieldSource::min_radius
+    pub conform_to_sphere: bool,
+}
+
+impl Default for ForceFieldSource {
+    fn default() -> Self {
+        // defaults to no force field (a mass of 0)
+        ForceFieldSource {
+            position: Vec3::new(0., 0., 0.),
+            min_radius: 0.1,
+            max_radius: 0.0,
+            mass: 0.,
+            force_exponent: 0.0,
+            conform_to_sphere: false,
+        }
+    }
+}
+
+impl ForceFieldSource {
+    /// Maximum number of sources in the force field.
+    pub const MAX_SOURCES: usize = 16;
+}
+
+/// A modifier to apply a force field to all particles each frame. The force
+/// field is made up of [`ForceFieldSource`]s.
+///
+/// The maximum number of sources is [`ForceFieldSource::MAX_SOURCES`].
+#[derive(Default, Clone, Copy)]
+pub struct ForceFieldModifier {
+    /// Array of force field sources.
+    ///
+    /// A source can be disabled by setting its [`mass`] to zero. In that case,
+    /// all other sources located after it in the array are automatically
+    /// disabled too.
+    ///
+    /// [`mass`]: ForceFieldSource::mass
+    pub sources: [ForceFieldSource; ForceFieldSource::MAX_SOURCES],
+}
+
+impl ForceFieldModifier {
+    /// Instantiate a [`ForceFieldModifier`] with a set of sources.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the number of sources exceeds [`MAX_SOURCES`].
+    ///
+    /// [`MAX_SOURCES`]: ForceFieldSource::MAX_SOURCES
+    pub fn new<T>(sources: T) -> Self
+    where
+        T: IntoIterator<Item = ForceFieldSource>,
+    {
+        let mut source_array = [ForceFieldSource::default(); ForceFieldSource::MAX_SOURCES];
+
+        for (index, p_attractor) in sources.into_iter().enumerate() {
+            if index > ForceFieldSource::MAX_SOURCES {
+                panic!(
+                    "Force field source count exceeded maximum of {}",
+                    ForceFieldSource::MAX_SOURCES
+                );
+            }
+            source_array[index] = p_attractor;
+        }
+
+        Self {
+            sources: source_array,
+        }
+    }
+
+    /// Overwrite the source at the given index.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `index` is greater than or equal to [`MAX_SOURCES`].
+    ///
+    /// [`MAX_SOURCES`]: ForceFieldSource::MAX_SOURCES
+    pub fn add_or_replace(&mut self, source: ForceFieldSource, index: usize) {
+        self.sources[index] = source;
+    }
+}
+
+impl UpdateModifier for ForceFieldModifier {
+    fn apply(&self, layout: &mut UpdateLayout) {
+        layout.force_field = self.sources;
+    }
+}
